backend/requirements.txt
openai
python-dotenv==1.0.1
pypdf==5.1.0
markdown-it-py==3.0.0
chromadb==0.5.5
chroma-hnswlib
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
fastapi==0.115.0
sse-starlette==2.1.0
uvicorn==0.30.6
pillow==10.4.0
numpy>=1.26
pydantic==2.9.2         
httpx==0.27.0            
streamlit==1.38.0      
